Corrections and extra bits
Total consumption write out to file did not work due to it being one value. Instead, I have added consumption as a method within the agent framework.
I have also defined __str__ so that the coords and store are printed when agents are printed. ( I can only figure out how to print it correctly for one agent eg, print(agent[i]) or print(agent[0]), but not for the whole list eg, print(agents) or print(agents[:]).